services/s3vibe/authproxy/requirements.txt
twisted==23.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
treq==23.11.0
requests==2.31.0

